Draft: move and rename DraftSelectPlane
